split components for better maintainability
