+++ src/styles/Post.css
@@ -1,0 +1,77 @@
+.Post {
+  width: 60%;
+  padding: 15px;
+}
+
+/* First Major Component Of Post */
+.PostInfo {
+  height: 100%;
+  padding: 5px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  /* Testing */
+  border: solid grey 2px;
+}
+
+.postInfo-main {
+  width: 100%;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+}
+.postInfo-description {
+  padding-top: 2px;
+}
+
+.postInfo-main-user {
+  padding-left: 25px;
+}
+
+.postInfo-main-details {
+  font-size: 15px;
+  margin-right: auto;
+  padding-left: 25px;
+}
+.details-main-username {
+  font-weight: bolder;
+}
+
+.postInfo-main-editPost {
+  padding-right: 15px;
+}
+
+.editPost-btn {
+  padding: 5px 10px 5px 10px;
+  border-radius: 30%;
+  cursor: pointer;
+}
+.editPost-btn:hover {
+  background-color: rgb(214, 215, 216);
+}
+
+/* SECOND MAJOR POST COMPONENT */
+
+.PostNavbar {
+  margin-top: 10px;
+  border-top: solid 2px grey;
+  border-bottom: solid 2px grey;
+
+  display: flex;
+  justify-content: space-around;
+}
+
+.postNavbar-item {
+  flex: 1;
+  display: block;
+
+  padding: 5px 0px 5px 0px;
+
+  text-align: center;
+  cursor: pointer;
+}
+.postNavbar-item:hover {
+  background-color: rgb(214, 215, 216);
+}
