+++ src/styles/PostContainer.css
@@ -3,51 +3,3 @@
   display: flex;
   justify-content: center;
 }
-
-.Post {
-  width: 60%;
-  outline: auto;
-}
-
-.postInfoContainer,
-.postNavbarContainer,
-.postCommentsContainer,
-.postAddCommentContainer {
-  height: 50px;
-  outline: auto;
-}
-
-/* First Major Component Of Post */
-.PostInfo {
-  height: 100%;
-
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-}
-
-.postInfo-user {
-  padding-left: 25px;
-}
-
-.postInfo-details {
-  font-size: 15px;
-  margin-right: auto;
-  padding-left: 25px;
-}
-.details-username {
-  font-weight: bolder;
-}
-
-.postInfo-editPost {
-  padding-right: 15px;
-}
-
-.editPost-btn {
-  padding: 5px 10px 5px 10px;
-  border-radius: 30%;
-  cursor: pointer;
-}
-.editPost-btn:hover {
-  background-color: rgb(214, 215, 216);
-}
